+++ app/indices/place_index.rb
@@ -1,35 +1,23 @@
-class Place < ActiveRecord::Base
+ThinkingSphinx::Index.define :place, with: :active_record, delta: ThinkingSphinx::Deltas::DelayedDelta do
+  indexes :name
+  indexes display_name
+  has place_type
 
-  acts_as_elastic_model
-  scope :load_for_index, -> { includes(:place_geometry) }
-  settings index: { number_of_shards: 1, analysis: ElasticModel::ANALYSIS } do
-    mappings(dynamic: true) do
-      indexes :geometry_geojson, type: "geo_shape"
-      indexes :location, type: "geo_point"
-      indexes :point_geojson, type: "geo_shape"
-      indexes :bbox_area, type: "double"
-      indexes :display_name, index_analyzer: "ascii_snowball_analyzer",
-        search_analyzer: "ascii_snowball_analyzer"
-      indexes :display_name_autocomplete, index_analyzer: "keyword_autocomplete_analyzer",
-        search_analyzer: "keyword_analyzer"
-    end
-  end
+  # HACK: TS doesn't seem to include attributes in the GROUP BY correctly
+  # for Postgres when using custom SQL attr definitions.  It may or may not
+  # be fixed in more up-to-date versions, but the issue has been raised:
+  # http://groups.google.com/group/thinking-sphinx/browse_thread/thread/e8397477b201d1e4
+  has :latitude, as: :fake_latitude
+  has :longitude, as: :fake_longitude
+  # END HACK
 
-  def as_indexed_json(options={})
-    preload_for_elastic_index
-    {
-      id: id,
-      name: name,
-      display_name: display_name,
-      display_name_autocomplete: display_name,
-      place_type: place_type,
-      bbox_area: bbox_area,
-      ancestor_place_ids: ancestor_place_ids,
-      geometry_geojson: place_geometry ?
-        ElasticModel.geom_geojson(place_geometry.geom) : nil,
-      location: ElasticModel.point_latlon(latitude, longitude),
-      point_geojson: ElasticModel.point_geojson(latitude, longitude)
-    }
-  end
+  # This is super brittle: the sphinx doc identifier here is based on
+  # ThinkingSphinx.unique_id_expression, which I can't get to work here, so
+  # if the number of indexed models changes this will break.
+  has "SELECT places.id * 12 + 1 AS id, regexp_split_to_table(id::text ||
+    (CASE WHEN ancestry IS NULL THEN '' ELSE '/' || ancestry END), '/')
+    AS place_ids FROM places", as: :place_ids, source: :query, type: :integer, facet: true, multi: true
 
+  has "RADIANS(latitude)", as: :latitude, type: :float
+  has "RADIANS(longitude)", as: :longitude, type: :float
 end
