+++ app/indices/project_index.rb
@@ -1,31 +1,11 @@
-class Project < ActiveRecord::Base
-
-  acts_as_elastic_model
-  scope :load_for_index, -> { includes(:place) }
-  settings index: { number_of_shards: 1, analysis: ElasticModel::ANALYSIS } do
-    mappings(dynamic: true) do
-      indexes :title, index_analyzer: "ascii_snowball_analyzer",
-        search_analyzer: "ascii_snowball_analyzer"
-      indexes :title_autocomplete, index_analyzer: "keyword_autocomplete_analyzer",
-        search_analyzer: "keyword_analyzer"
-      indexes :description, index_analyzer: "ascii_snowball_analyzer",
-        search_analyzer: "ascii_snowball_analyzer"
-      indexes :location, type: "geo_point"
-      indexes :geojson, type: "geo_shape"
-    end
-  end
-
-  def as_indexed_json(options={})
-    preload_for_elastic_index
-    {
-      id: id,
-      title: title,
-      title_autocomplete: title,
-      description: description,
-      ancestor_place_ids: place ? place.ancestor_place_ids : nil,
-      location: ElasticModel.point_latlon(latitude, longitude),
-      geojson: ElasticModel.point_geojson(latitude, longitude)
-    }
-  end
-
+ThinkingSphinx::Index.define :project, with: :active_record, delta: ThinkingSphinx::Deltas::DelayedDelta do
+  indexes :title
+  indexes :description
+  # This is super brittle: the sphinx doc identifier here is based on
+  # ThinkingSphinx.unique_id_expression, which I can't get to work here, so
+  # if the number of indexed models changes this will break.
+  has "SELECT projects.id * 12 + 2 AS id, regexp_split_to_table(projects.place_id::text ||
+    (CASE WHEN ancestry IS NULL THEN '' ELSE '/' || ancestry END), '/')
+    AS place_id FROM projects JOIN places ON place_id = places.id",
+    as: :place_ids, source: :query, facet: true, type: :integer, multi: true
 end
