+++ spec/lib/elastic_model/indices/place_index_spec.rb
@@ -1,0 +1,9 @@
+require "spec_helper"
+
+describe "Place Index" do
+  it "as_indexed_json should return a hash" do
+    p = Place.make!
+    json = p.as_indexed_json
+    expect( json ).to be_a Hash
+  end
+end
