+++ spec/lib/elastic_model/indices/project_index_spec.rb
@@ -1,0 +1,9 @@
+require "spec_helper"
+
+describe "Project Index" do
+  it "as_indexed_json should return a hash" do
+    p = Project.make!
+    json = p.as_indexed_json
+    expect( json ).to be_a Hash
+  end
+end
